+++ CHANGELOG.md
@@ -5,5 +5,10 @@
 ## 1.0.1
 
 -Fixing the convertTimeStampToStr(Map<String, dynamic> myMap) to be able to convert TimeStamp values in nested Map
--You don't need to provide the fields keys you want to convert to TimeStamp inside #convertStrToTimeStamp(Map<String, dynamic> myMap) function .
+-You don't need to provide the fields keys you want to convert to TimeStamp inside convertStrToTimeStamp(Map<String, dynamic> myMap) function .
+
+## 1.0.2
+
+ -Adding containsTimeStamp() method to check if a map contains any TimeStamp value
+ -Adding containsISO8601String() method to check if a map contains any ISO 8601 String value